+++ lib/src/serializer.dart
@@ -1,16 +1,34 @@
-import 'php_serializer.dart';
+import '../php_serializer.dart';
 
 /// Takes an object and converts it into a [String] which could be deserialized
-/// by Php via its function `deserialize()`.
+/// by Php via its function `unserialize()`.
 ///
 /// Only fundamental Objects are recognized, every other object requires
 /// additional information via a [List] of [PhpSerializationObjectInformation]
 /// as the second argument.
+/// Alternatively the third argument can be used to specify different behavior
+/// in case there is no matching [PhpSerializationObjectInformation] for
+/// classes.
 String phpSerialize(dynamic serializeMe,
-    [List<PhpSerializationObjectInformation>? knownClasses]) {
-  return _Serializer.parse(serializeMe, knownClasses ?? []);
+    [List<PhpSerializationObjectInformation>? knownClasses,
+    NoMatchingObjectSerializationInformation? fallbackObjectSerialization]) {
+  final serializer = _Serializer(
+      knownClasses ?? [],
+      fallbackObjectSerialization ??
+          ThrowExceptionOnMissingSerializationInformation());
+
+  return serializer.parse(serializeMe);
 }
 
+/// Implement this interface to customize the action when there is not matching
+/// information for serialization.
+/// {@category Fallback Behavior}
+abstract class NoMatchingObjectSerializationInformation {
+  /// Return a [PhpSerializationObjectInformation] for the requested class
+  PhpSerializationObjectInformation handleSerialization(Type objectType);
+}
+
+/// Will be thrown in case [phpSerialize] fails for some reason
 abstract class SerializationException implements Exception {
   final String message;
 
@@ -26,7 +44,7 @@ class ObjectWithoutSerializationInformationFound
     extends SerializationException {
   ObjectWithoutSerializationInformationFound(Type objectType)
       : super(
-            'An object of type ${objectType.toString()} couldn\'t be serialized, since no serialiazion-information was provided!');
+            'An object of type ${objectType.toString()} couldn\'t be serialized, since no serialization-information was provided!');
 }
 
 ///Serialization of an object failed because the user-defined converter
@@ -41,8 +59,12 @@ class CustomSerializationFailed extends SerializationException {
 }
 
 class _Serializer {
-  static String parse(dynamic rawInput,
-      List<PhpSerializationObjectInformation> objectInformation) {
+  final List<PhpSerializationObjectInformation> _objectInformation;
+  final NoMatchingObjectSerializationInformation _fallbackObjectSerialization;
+
+  _Serializer(this._objectInformation, this._fallbackObjectSerialization);
+
+  String parse(dynamic rawInput) {
     switch (rawInput.runtimeType) {
       case String:
         return _parseString(rawInput);
@@ -58,74 +80,70 @@ class _Serializer {
     }
 
     if (rawInput is List) {
-      return _parseList(rawInput, objectInformation);
+      return _parseList(rawInput);
     }
 
     if (rawInput is Map) {
-      return _parseMap(rawInput, objectInformation);
+      return _parseMap(rawInput);
     }
 
-    return _parseSerializableClass(rawInput, objectInformation);
+    return _parseSerializableClass(rawInput);
   }
 
-  static String _parseString(String inputString) {
+  String _parseString(String inputString) {
     return 's:${inputString.length}:"$inputString";';
   }
 
-  static String _parseInt(int inputInt) {
+  String _parseInt(int inputInt) {
     return 'i:$inputInt;';
   }
 
-  static String _parseDouble(double inputDouble) {
+  String _parseDouble(double inputDouble) {
     if (inputDouble.floorToDouble() == inputDouble) {
       return 'd:${inputDouble.floor()};';
     }
     return 'd:$inputDouble;';
   }
 
-  static String _parseList(List<dynamic> inputList,
-      List<PhpSerializationObjectInformation> objectInformation) {
+  String _parseList(List<dynamic> inputList) {
     final sb = StringBuffer('a:${inputList.length}:{');
     for (var i = 0; i != inputList.length; ++i) {
       sb.write('i:$i;');
-      sb.write(parse(inputList[i], objectInformation));
+      sb.write(parse(inputList[i]));
     }
 
     sb.write('}');
     return sb.toString();
   }
 
-  static String _parseMap(Map<dynamic, dynamic> inputMap,
-      List<PhpSerializationObjectInformation> objectInformation,
+  String _parseMap(Map<dynamic, dynamic> inputMap,
       {bool prependArrayIdentifierAndSize = true}) {
     final sb = StringBuffer();
     if (prependArrayIdentifierAndSize) sb.write('a:${inputMap.length}:');
     sb.write('{');
 
     inputMap.forEach((key, value) {
-      sb.write(
-          '${parse(key, objectInformation)}${parse(value, objectInformation)}');
+      sb.write('${parse(key)}${parse(value)}');
     });
 
     sb.write('}');
     return sb.toString();
   }
 
-  static String _parseSerializableClass(Object inputClass,
-      List<PhpSerializationObjectInformation> objectInformation) {
-    final matchingObject = objectInformation.firstWhere(
+  String _parseSerializableClass(Object inputClass) {
+    final matchingObject = _objectInformation.firstWhere(
         (element) => element.typeOf == inputClass.runtimeType,
-        orElse: () => throw ObjectWithoutSerializationInformationFound(
-            inputClass.runtimeType));
+        orElse: () => _fallbackObjectSerialization
+            .handleSerialization(inputClass.runtimeType));
     final className = matchingObject.serializedClassName;
-    late final intermediateMap;
+    late final Map<String, dynamic> intermediateMap;
     try {
       intermediateMap = matchingObject.dataExtractor(inputClass);
     } catch (e) {
       throw CustomSerializationFailed(matchingObject.typeOf, e);
     }
-    final serializer = _parseMap(intermediateMap, objectInformation,
-        prependArrayIdentifierAndSize: false);
-    return 'O:${className.length}:"$className":${intermediateMap.length}:$serializer';
+    final serializedMap =
+        _parseMap(intermediateMap, prependArrayIdentifierAndSize: false);
+    return 'O:${className.length}:"$className":${intermediateMap.length}:$serializedMap';
   }
 }