+++ test/deserialize_symfony_example_test.dart
@@ -1,4 +1,3 @@
-import 'package:php_serializer/deserializer.dart';
 import 'package:php_serializer/php_serializer.dart';
 import 'package:test/test.dart';