+++ test/deserializer_test.dart
@@ -1,5 +1,5 @@
-import 'package:php_serializer/deserializer.dart';
 import 'package:test/test.dart';
+
 import 'package:php_serializer/php_serializer.dart';
 import 'testClasses.dart';